requirements.txt
asgiref==3.4.1
Django==3.2.7
djangorestframework==3.12.4
psycopg2==2.8.1
pytz==2021.1
sqlparse==0.4.2
uWSGI>2.0.18,<2.1
six==1.16
